refactor: split sources and update header
