+++ include/push_swap.h
@@ -4,7 +4,6 @@
 # include <stdlib.h>
 # include <limits.h>
 
-/* ===================== STRUCTS ===================== */
 typedef struct s_node
 {
     int             value;
@@ -12,11 +11,11 @@ typedef struct s_node
     struct s_node   *next;
 } t_node;
 
-/* ===================== PARSING ===================== */
 t_node  *parse_input(int argc, char **argv);
 char    **split_args(int argc, char **argv);
 int     count_all_tokens(int argc, char **argv);
-int     has_duplicates(t_node *stack);
+int is_valid_number(const char *str);
+int add_node(t_node **stack, const char *str);
 char    **ft_split(const char *s, char c);
 char    *ft_strdup(const char *s);
 void    free_split(char **tokens);
@@ -24,29 +23,32 @@ int     ft_atoi_push_swap(const char *str, int *error_flag);
 void    free_tokens(char **tokens, int count);
 void    free_stack(t_node **stack);
 
-/* ==================== MOVIMIENTOS ================== */
-void sa(t_node **a); // swap a
-void sb(t_node **b); // swap b
-void ss(t_node **a, t_node **b); // swap a y b
+void    sa(t_node **a); /* swap a */
+void    sb(t_node **b); /* swap b */
+void    ss(t_node **a, t_node **b); /* swap a and b */
 
-void pa(t_node **a, t_node **b); // push a (b → a)
-void pb(t_node **a, t_node **b); // push b (a → b)
+void    pa(t_node **a, t_node **b); /* push b to a */
+void    pb(t_node **a, t_node **b); /* push a to b */
 
-void ra(t_node **a); // rotate a
-void rb(t_node **b); // rotate b
-void rr (t_node **a, t_node **b); //rotate a y b
+void    ra(t_node **a); /* rotate a */
+void    rb(t_node **b); /* rotate b */
+void    rr(t_node **a, t_node **b); /* rotate a and b */
 
-void rra(t_node **a); // reverse rotate a
-void rrb(t_node **b); // reverse rotate b
-void rrr(t_node **a, t_node **b); // reverse rotate a y b
+void    rra(t_node **a); /* reverse rotate a */
+void    rrb(t_node **b); /* reverse rotate b */
+void    rrr(t_node **a, t_node **b); /* reverse rotate a and b */
 
-/* ==================== UTILS ======================== */
 void    swap_nodes(t_node **stack);
 void    push_node(t_node **dst, t_node **src);
 void    rotate_nodes(t_node **stack);
 void    reverse_rotate_nodes(t_node **stack);
 
-/* ==================== ALGORITMO ==================== */
+int     stack_size(t_node *stack);
+void    assign_indexes(t_node *stack);
+void    sort_three(t_node **a);
+void    sort_five(t_node **a, t_node **b);
+void    chunk_sort(t_node **a, t_node **b, int chunks);
+
 int     is_sorted(t_node *stack);
 void    sort_stack(t_node **a, t_node **b);
