+++ src/algorithm/stack_utils.c
@@ -1,0 +1,14 @@
+#include "push_swap.h"
+
+int stack_size(t_node *stack)
+{
+    int size;
+
+    size = 0;
+    while (stack)
+    {
+        size++;
+        stack = stack->next;
+    }
+    return size;
+}
