+++ src/movements/push_ops.c
@@ -1,0 +1,18 @@
+#include "push_swap.h"
+#include <unistd.h>
+
+void pa(t_node **a, t_node **b)
+{
+    if (!b || !*b)
+        return;
+    push_node(a, b);
+    write(1, "pa\n", 3);
+}
+
+void pb(t_node **a, t_node **b)
+{
+    if (!a || !*a)
+        return;
+    push_node(b, a);
+    write(1, "pb\n", 3);
+}
